+++ src/main/java/com/example/OrderManagementSystem/repository/ContractRepository.java
@@ -7,7 +7,7 @@ import java.util.*;
 public class ContractRepository {
 
     private final Map<Long, Contract> contractStore = new HashMap<>();
-    private Long nextId = 1L;
+    private int nextId = 1;
 
     public Contract save(Contract contract) {
         if (contract.getId() == null) {
